moving everything to ruler.js
