+++ css/herosystem6e.css
@@ -706,7 +706,7 @@ td {
   width: 200px;
   font-size: 14px;
   position: absolute;
-  top: 70px;
+  top: 120px;
   left: 15px;
   width: 200px;
 }
